# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FinSense AI API",
  "description": "Financial analytics, portfolio optimisation, ML forecasting and RAG research API. Decision support only - not investment advice.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/dashboard": {
   "get": {
    "tags": [
     "insights"
    ],
    "summary": "Dashboard",
    "operationId": "dashboard_api_v1_dashboard_get",
    "parameters": [
     {
      "name": "portfolio_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Portfolio Id"
      }
     },
     {
      "name": "fund_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Fund Id"
      }
     },
     {
      "name": "period",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "1y",
        "3y",
        "5y",
        "max"
       ],
       "type": "string",
       "default": "3y",
       "title": "Period"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Dashboard Api V1 Dashboard Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/simulate/sip": {
   "post": {
    "tags": [
     "insights"
    ],
    "summary": "Simulate Sip",
    "operationId": "simulate_sip_api_v1_simulate_sip_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SipRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Simulate Sip Api V1 Simulate Sip Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/simulate/monte-carlo": {
   "post": {
    "tags": [
     "insights"
    ],
    "summary": "Simulate Monte Carlo",
    "operationId": "simulate_monte_carlo_api_v1_simulate_monte_carlo_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MonteCarloRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Simulate Monte Carlo Api V1 Simulate Monte Carlo Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/simulate/market-shock": {
   "post": {
    "tags": [
     "insights"
    ],
    "summary": "Simulate Market Shock",
    "operationId": "simulate_market_shock_api_v1_simulate_market_shock_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ShockRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Simulate Market Shock Api V1 Simulate Market Shock Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/simulate/stress": {
   "post": {
    "tags": [
     "insights"
    ],
    "summary": "Simulate Stress",
    "operationId": "simulate_stress_api_v1_simulate_stress_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StressRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Simulate Stress Api V1 Simulate Stress Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/simulate/allocation-change": {
   "post": {
    "tags": [
     "insights"
    ],
    "summary": "Simulate Allocation Change",
    "operationId": "simulate_allocation_change_api_v1_simulate_allocation_change_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AllocationChangeRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Simulate Allocation Change Api V1 Simulate Allocation Change Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/ml/forecast": {
   "post": {
    "tags": [
     "insights"
    ],
    "summary": "Ml Forecast",
    "operationId": "ml_forecast_api_v1_ml_forecast_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ForecastRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Ml Forecast Api V1 Ml Forecast Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ml/anomalies": {
   "post": {
    "tags": [
     "insights"
    ],
    "summary": "Ml Anomalies",
    "operationId": "ml_anomalies_api_v1_ml_anomalies_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AnomalyRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Ml Anomalies Api V1 Ml Anomalies Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ml/models": {
   "get": {
    "tags": [
     "insights"
    ],
    "summary": "Ml Models",
    "operationId": "ml_models_api_v1_ml_models_get",
    "parameters": [
     {
      "name": "fund_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Fund Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Ml Models Api V1 Ml Models Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AllocationChangeRequest": {
    "properties": {
     "fund_ids": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "maxItems": 20,
      "minItems": 1,
      "title": "Fund Ids"
     },
     "current_weights_pct": {
      "items": {
       "type": "number"
      },
      "type": "array",
      "title": "Current Weights Pct"
     },
     "proposed_weights_pct": {
      "items": {
       "type": "number"
      },
      "type": "array",
      "title": "Proposed Weights Pct"
     },
     "lookback": {
      "type": "string",
      "enum": [
       "1y",
       "3y",
       "5y",
       "max"
      ],
      "title": "Lookback",
      "default": "3y"
     }
    },
    "type": "object",
    "required": [
     "fund_ids",
     "current_weights_pct",
     "proposed_weights_pct"
    ],
    "title": "AllocationChangeRequest"
   },
...
   "AnomalyRequest": {
    "properties": {
     "fund_id": {
      "type": "integer",
      "title": "Fund Id"
     },
     "contamination": {
      "type": "number",
      "maximum": 0.1,
      "minimum": 0.001,
      "title": "Contamination",
      "default": 0.01
     }
    },
    "type": "object",
    "required": [
     "fund_id"
    ],
    "title": "AnomalyRequest"
   },
...
   "ForecastRequest": {
    "properties": {
     "fund_id": {
      "type": "integer",
      "title": "Fund Id"
     },
     "horizon_days": {
      "type": "integer",
      "enum": [
       5,
       21,
       63
      ],
      "title": "Horizon Days",
      "default": 21
     },
     "model_type": {
      "type": "string",
      "enum": [
       "ridge",
       "random_forest",
       "historical_mean"
      ],
      "title": "Model Type",
      "default": "random_forest"
     },
     "explain": {
      "type": "boolean",
      "title": "Explain",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "fund_id"
    ],
    "title": "ForecastRequest"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MonteCarloRequest": {
    "properties": {
     "monthly_amount": {
      "type": "number",
      "maximum": 100000000.0,
      "minimum": 0.0,
      "title": "Monthly Amount"
     },
     "years": {
      "type": "integer",
      "maximum": 50.0,
      "minimum": 1.0,
      "title": "Years"
     },
     "annual_return": {
      "type": "number",
      "exclusiveMaximum": 1.0,
      "exclusiveMinimum": -0.5,
      "title": "Annual Return"
     },
     "lumpsum": {
      "type": "number",
      "maximum": 10000000000.0,
      "minimum": 0.0,
      "title": "Lumpsum",
      "default": 0
     },
     "annual_step_up": {
      "type": "number",
      "maximum": 0.5,
      "minimum": 0.0,
      "title": "Annual Step Up",
      "default": 0
     },
     "annual_volatility": {
      "type": "number",
      "exclusiveMaximum": 1.5,
      "minimum": 0.0,
      "title": "Annual Volatility"
     },
     "paths": {
      "type": "integer",
      "maximum": 20000.0,
      "minimum": 100.0,
      "title": "Paths",
      "default": 2000
     },
     "seed": {
      "type": "integer",
      "maximum": 2147483647.0,
      "minimum": 0.0,
      "title": "Seed",
      "default": 7
     }
    },
    "type": "object",
    "required": [
     "monthly_amount",
     "years",
     "annual_return",
     "annual_volatility"
    ],
    "title": "MonteCarloRequest"
   },
...
   "ShockRequest": {
    "properties": {
     "portfolio_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Portfolio Id"
     },
     "fund_ids": {
      "anyOf": [
       {
        "items": {
         "type": "integer"
        },
        "type": "array",
        "maxItems": 20,
        "minItems": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Fund Ids"
     },
     "weights_pct": {
      "anyOf": [
       {
        "items": {
         "type": "number"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Weights Pct"
     },
     "lookback": {
      "type": "string",
      "enum": [
       "1y",
       "3y",
       "5y",
       "max"
      ],
      "title": "Lookback",
      "default": "3y"
     },
     "market_move": {
      "type": "number",
      "maximum": 0.5,
      "minimum": -0.9,
      "title": "Market Move"
     },
     "portfolio_value": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Portfolio Value"
     }
    },
    "type": "object",
    "required": [
     "market_move"
    ],
    "title": "ShockRequest"
   },
   "SipRequest": {
    "properties": {
     "monthly_amount": {
      "type": "number",
      "maximum": 100000000.0,
      "minimum": 0.0,
      "title": "Monthly Amount"
     },
     "years": {
      "type": "integer",
      "maximum": 50.0,
      "minimum": 1.0,
      "title": "Years"
     },
     "annual_return": {
      "type": "number",
      "exclusiveMaximum": 1.0,
      "exclusiveMinimum": -0.5,
      "title": "Annual Return"
     },
     "lumpsum": {
      "type": "number",
      "maximum": 10000000000.0,
      "minimum": 0.0,
      "title": "Lumpsum",
      "default": 0
     },
     "annual_step_up": {
      "type": "number",
      "maximum": 0.5,
      "minimum": 0.0,
      "title": "Annual Step Up",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "monthly_amount",
     "years",
     "annual_return"
    ],
    "title": "SipRequest"
   },
   "StressRequest": {
    "properties": {
     "portfolio_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Portfolio Id"
     },
     "fund_ids": {
      "anyOf": [
       {
        "items": {
         "type": "integer"
        },
        "type": "array",
        "maxItems": 20,
        "minItems": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Fund Ids"
     },
     "weights_pct": {
      "anyOf": [
       {
        "items": {
         "type": "number"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Weights Pct"
     },
     "lookback": {
      "type": "string",
      "enum": [
       "1y",
       "3y",
       "5y",
       "max"
      ],
      "title": "Lookback",
      "default": "3y"
     },
     "volatility_multiplier": {
      "type": "number",
      "maximum": 5.0,
      "minimum": 0.1,
      "title": "Volatility Multiplier",
      "default": 1.0
     },
     "correlation_override": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 1.0,
        "minimum": -1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Correlation Override"
     },
     "return_shift": {
      "type": "number",
      "maximum": 0.5,
      "minimum": -0.5,
      "title": "Return Shift",
      "default": 0.0
     },
     "horizon_days": {
      "type": "integer",
      "maximum": 252.0,
      "minimum": 1.0,
      "title": "Horizon Days",
      "default": 1
     },
     "confidence": {
      "type": "number",
      "maximum": 0.995,
      "minimum": 0.9,
      "title": "Confidence",
      "default": 0.95
     }
    },
    "type": "object",
    "title": "StressRequest"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}